erdos_graph::logger::core

Trait Logger

pub trait Logger:
    Send
    + Sync
    + 'static {
    // Required method
    fn log(&self, level: LogLevel, message: &str);

    // Provided methods
    fn flush(&self) { ... }
    fn min_level(&self) -> LogLevel { ... }
    fn trace(&self, message: &str) { ... }
    fn debug(&self, message: &str) { ... }
    fn info(&self, message: &str) { ... }
    fn warn(&self, message: &str) { ... }
    fn error(&self, message: &str) { ... }
}

Minimal logger interface used throughout the project.

Implementors should be cheap to clone and thread-safe if used in multithreaded contexts. For stubs, implement log and optionally override convenience methods. TODO: consider adding Send + Sync bounds once we finalize the threading model. The logger trait used across the project.

Implementations must be Send + Sync so they can be shared across threads when the global facade is used.

Required Methods

fn log(&self, level: LogLevel, message: &str)

Emit a log record at the given level.

Provided Methods

fn flush(&self)

Flush any buffered records.

fn min_level(&self) -> LogLevel

Get the current minimum log level.

fn trace(&self, message: &str)

Convenience: trace level.

fn debug(&self, message: &str)

Convenience: debug level.

fn info(&self, message: &str)

Convenience: info level.

fn warn(&self, message: &str)

Convenience: warn level.

fn error(&self, message: &str)

Convenience: error level.